app/requirements.txt
fastapi==0.65.2
uvicorn==0.14.0
typing==3.7.4.3
SQLAlchemy==1.4.19
passlib==1.7.4
bcrypt==3.2.0
python-jose==3.3.0
alembic
psycopg2
requests
graphene==2.1.8
graphene-sqlalchemy==2.3.0
black==21.6b0
python-dotenv==0.18.0